move stuff out of PostgresqlDatabaseQueries
