+++ site/app/libraries/database/PostgresqlDatabaseQueries.php
@@ -15,11 +15,6 @@ use app\models\SimpleStat;
 
 class PostgresqlDatabaseQueries extends DatabaseQueries{
 
-    public function getSubmittyUser($user_id) {
-        $this->submitty_db->query("SELECT * FROM users WHERE user_id=?", array($user_id));
-        return ($this->submitty_db->rowCount() > 0) ? new User($this->core, $this->submitty_db->row()) : null;
-    }
-
     public function getUserById($user_id) {
         $this->course_db->query("
 SELECT u.*, sr.grading_registration_sections
@@ -127,24 +122,6 @@ WHERE semester=? AND course=? AND user_id=?", $params);
         $params = array($user->getRotatingSection(), $user->getId());
         $this->course_db->query("UPDATE users SET rotating_section=? WHERE user_id=?", $params);
         $this->updateGradingRegistration($user->getId(), $user->getGroup(), $user->getGradingRegistrationSections());
-    }
-
-    public function updateGradingRegistration($user_id, $user_group, $sections) {
-        $this->course_db->query("DELETE FROM grading_registration WHERE user_id=?", array($user_id));
-        if ($user_group < 4) {
-            foreach ($sections as $section) {
-                $this->course_db->query("
-    INSERT INTO grading_registration (user_id, sections_registration_id) VALUES(?, ?)", array($user_id, $section));
-            }
-        }
-    }
-
-    public function getAllGradeables($user_id = null) {
-        return $this->getGradeables(null, $user_id);
-    }
-
-    public function getGradeable($g_id = null, $user_id = null) {
-        return $this->getGradeables($g_id, $user_id)[0];
     }
 
     public function getGradeablesIterator($g_ids = null, $user_ids = null, $section_key="registration_section", $sort_key="u.user_id", $g_type = null) {
@@ -422,112 +399,17 @@ LEFT JOIN (
         }
         if (count($where) > 0) {
             $query .= "
-WHERE ".implode(" AND ", $where);
+WHERE " . implode(" AND ", $where);
         }
         if ($user_ids !== null) {
             $query .= "
 ORDER BY u.{$section_key}, {$sort_key}";
         }
 
-        return $this->course_db->queryIterator($query, $params, function($row) {
+        return $this->course_db->queryIterator($query, $params, function ($row) {
             $user = (isset($row['user_id']) && $row['user_id'] !== null) ? new User($this->core, $row) : null;
             return new Gradeable($this->core, $row, $user);
         });
-    }
-
-    /*
-     * TODO:
-     * This should take in for:
-     *  gradeable: [string] or [array] which then maps that into a where clause (g_id = string) OR (g_id IN (?, ?))
-     *  users: [string] or [array] which then maps that into a where clause as well as adding in additional
-     *      components for the SELECT cause and in the FROM clause (don't need gradeable_data if this is null, etc.)
-     *  section_key:
-     */
-    public function getGradeables($g_ids = null, $user_ids = null, $section_key="registration_section", $sort_key="u.user_id", $g_type = null) {
-        $return = array();
-        foreach ($this->getGradeablesIterator($g_ids, $user_ids, $section_key, $sort_key, $g_type) as $key => $row) {
-            $return[] = $row;
-        }
-
-        return $return;
-    }
-
-    public function getGradeableComponents($g_id, $gd_id=null) {
-        $this->course_db->query("
-SELECT gcd.*, gc.*
-FROM gradeable_component AS gc
-LEFT JOIN (
-  SELECT *
-  FROM gradeable_component_data
-  WHERE gd_id = ?
-) as gcd ON gc.gc_id = gcd.gc_id
-WHERE gc.g_id=?
-", array($gd_id, $g_id));
-
-        $return = array();
-        foreach ($this->course_db->rows() as $row) {
-            $return[$row['gc_id']] = new GradeableComponent($this->core, $row);
-        }
-        return $return;
-    }
-
-    public function getGradeableComponentsMarks($gc_id) {
-      $this->course_db->query("
-SELECT *
-FROM gradeable_component_mark
-WHERE gc_id=?
-ORDER BY gcm_order ASC
-", array($gc_id));
-      $return = array();
-        foreach ($this->course_db->rows() as $row) {
-            $return[$row['gcm_id']] = new GradeableComponentMark($this->core, $row);
-        }
-        return $return;
-    }
-
-// This has to be updated to also load components for each version
-    public function getGradeableVersions($g_id, $user_id, $team_id, $due_date) {
-        if ($user_id === null) {
-            $this->course_db->query("
-SELECT egd.*, egv.active_version = egd.g_version as active_version
-FROM electronic_gradeable_data AS egd
-LEFT JOIN (
-  SELECT *
-  FROM electronic_gradeable_version
-) AS egv ON egv.active_version = egd.g_version AND egv.team_id = egd.team_id AND egv.g_id = egd.g_id
-WHERE egd.g_id=? AND egd.team_id=?
-ORDER BY egd.g_version", array($g_id, $team_id));
-        }
-        else {
-            $this->course_db->query("
-SELECT egd.*, egv.active_version = egd.g_version as active_version
-FROM electronic_gradeable_data AS egd
-LEFT JOIN (
-  SELECT *
-  FROM electronic_gradeable_version
-) AS egv ON egv.active_version = egd.g_version AND egv.user_id = egd.user_id AND egv.g_id = egd.g_id
-WHERE egd.g_id=? AND egd.user_id=?
-ORDER BY egd.g_version", array($g_id, $user_id));
-        }
-        
-        $return = array();
-        foreach ($this->course_db->rows() as $row) {
-            $row['submission_time'] = new \DateTime($row['submission_time'], $this->core->getConfig()->getTimezone());
-            $return[$row['g_version']] = new GradeableVersion($this->core, $row, $due_date);
-        }
-
-        return $return;
-    }
-
-    // Moved from class LateDaysCalculation on port from TAGrading server.  May want to incorporate late day information into gradeable object rather than having a separate query 
-    public function getLateDayUpdates($user_id) {
-        if($user_id != null) {
-          $this->course_db->query("SELECT * FROM late_days WHERE user_id=?", array($user_id));
-        }
-        else {
-          $this->course_db->query("SELECT * FROM late_days");
-        }
-        return $this->course_db->rows();
     }
     
     // Moved from class LateDaysCalculation on port from TAGrading server.  May want to incorporate late day information into gradeable object rather than having a separate query
@@ -597,72 +479,6 @@ ORDER BY egd.g_version", array($g_id, $user_id));
         }
         $this->course_db->query($query, $params);
         return $this->course_db->rows();
-    }
-
-    public function getUsersByRegistrationSections($sections) {
-        $return = array();
-        if (count($sections) > 0) {
-            $query = implode(",", array_fill(0, count($sections), "?"));
-            $this->course_db->query("SELECT * FROM users WHERE registration_section IN ({$query}) ORDER BY registration_section", $sections);
-            foreach ($this->course_db->rows() as $row) {
-                $return[] = new User($this->core, $row);
-            }
-        }
-        return $return;
-    }
-
-    public function getTotalUserCountByGradingSections($sections, $section_key) {
-        $return = array();
-        $params = array();
-        $where = "";
-        if (count($sections) > 0) {
-            $where = "WHERE {$section_key} IN (".implode(",", array_fill(0, count($sections), "?")).")";
-            $params = $sections;
-        }
-        $this->course_db->query("
-SELECT count(*) as cnt, {$section_key}
-FROM users 
-{$where}
-GROUP BY {$section_key} 
-ORDER BY {$section_key}", $params);
-        foreach ($this->course_db->rows() as $row) {
-            if ($row[$section_key] === null) {
-                $row[$section_key] = "NULL";
-            }
-            $return[$row[$section_key]] = intval($row['cnt']);
-        }
-        return $return;
-    }
-
-    public function getTotalComponentCount($g_id) {
-        $this->course_db->query("SELECT count(*) AS cnt FROM gradeable_component WHERE g_id=?", array($g_id));
-        return intval($this->course_db->row()['cnt']);
-    }
-
-    public function getGradedComponentsCountByGradingSections($g_id, $sections, $section_key) {
-        $return = array();
-        $params = array($g_id);
-        $where = "";
-        if (count($sections) > 0) {
-            $where = "WHERE {$section_key} IN (".implode(",", array_fill(0, count($sections), "?")).")";
-            $params = array_merge($params, $sections);
-        }
-        $this->course_db->query("
-SELECT count(u.*) as cnt, u.{$section_key}
-FROM(
-  SELECT * FROM gradeable_component_data AS gcd LEFT JOIN gradeable_data AS gd ON gcd.gd_id = gd.gd_id WHERE g_id=?
-) AS gd 
-INNER JOIN users AS u ON u.user_id = gd.gd_user_id
-{$where}
-GROUP BY u.{$section_key}
-ORDER BY u.{$section_key}", $params);
-        foreach ($this->course_db->rows() as $row) {
-            if ($row[$section_key] === null) {
-                $row[$section_key] = "NULL";
-            }
-            $return[$row[$section_key]] = intval($row['cnt']);
-        }
-        return $return;
     }
 
     public function getAverageComponentScores($g_id) {
@@ -741,111 +557,6 @@ SELECT round((AVG(g_score) + AVG(autograding)),2) AS avg_score, round(stddev_pop
         return new SimpleStat($this->core, $this->course_db->rows()[0]);
     }
 
-
-    public function getGradersForRegistrationSections($sections) {
-        $return = array();
-        $params = array();
-        $where = "";
-        if (count($sections) > 0) {
-            $where = "WHERE sections_registration_id IN (" . implode(",", array_fill(0, count($sections), "?")) . ")";
-            $params = $sections;
-        }
-        $this->course_db->query("
-SELECT g.*, u.* 
-FROM grading_registration AS g
-LEFT JOIN (
-  SELECT *
-  FROM users
-) AS u ON u.user_id = g.user_id
-{$where}
-ORDER BY g.sections_registration_id, g.user_id", $params);
-        $user_store = array();
-        foreach ($this->course_db->rows() as $row) {
-            if ($row['sections_registration_id'] === null) {
-                $row['sections_registration_id'] = "NULL";
-            }
-
-            if (!isset($return[$row['sections_registration_id']])) {
-                $return[$row['sections_registration_id']] = array();
-            }
-
-            if (!isset($user_store[$row['user_id']])) {
-                $user_store[$row['user_id']] = new User($this->core, $row);
-            }
-            $return[$row['sections_registration_id']][] = $user_store[$row['user_id']];
-        }
-        return $return;
-    }
-
-    public function getGradersForRotatingSections($g_id, $sections) {
-        $return = array();
-        $params = array($g_id);
-        $where = "";
-        if (count($sections) > 0) {
-            $where = " AND sections_rotating_id IN (" . implode(",", array_fill(0, count($sections), "?")) . ")";
-            $params = array_merge($params, $sections);
-        }
-        $this->course_db->query("
-SELECT g.*, u.* 
-FROM grading_rotating AS g
-LEFT JOIN (
-  SELECT *
-  FROM users
-) AS u ON u.user_id = g.user_id
-WHERE g.g_id=? {$where}
-ORDER BY g.sections_rotating_id, g.user_id", $params);
-        $user_store = array();
-        foreach ($this->course_db->rows() as $row) {
-            if ($row['sections_rotating_id'] === null) {
-                $row['sections_rotating_id'] = "NULL";
-            }
-            if (!isset($return[$row['sections_rotating_id']])) {
-                $return[$row['sections_rotating_id']] = array();
-            }
-            if (!isset($user_store[$row['user_id']])) {
-                $user_store[$row['user_id']] = new User($this->core, $row);
-            }
-            $return[$row['sections_rotating_id']][] = $user_store[$row['user_id']];
-        }
-        return $return;
-    }
-
-    public function getRotatingSectionsForGradeableAndUser($g_id, $user) {
-        $this->course_db->query("SELECT sections_rotating_id FROM grading_rotating WHERE g_id=? AND user_id=?", array($g_id, $user));
-        $return = array();
-        foreach ($this->course_db->rows() as $row) {
-            $return[] = $row['sections_rotating_id'];
-        }
-        return $return;
-    }
-
-    public function getUsersByRotatingSections($sections) {
-        $return = array();
-        if (count($sections) > 0) {
-            $query = implode(",", array_fill(0, count($sections), "?"));
-            $this->course_db->query("SELECT * FROM users WHERE rotating_section IN ({$query}) ORDER BY rotating_section", $sections);
-            foreach ($this->course_db->rows() as $row) {
-                $return[] = new User($this->core, $row);
-            }
-        }
-        return $return;
-    }
-
-    public function getRegistrationSections() {
-        $this->course_db->query("SELECT * FROM sections_registration ORDER BY sections_registration_id");
-        return $this->course_db->rows();
-    }
-
-    public function getRotatingSections() {
-        $this->course_db->query("SELECT * FROM sections_rotating ORDER BY sections_rotating_id");
-        return $this->course_db->rows();
-    }
-
-    public function getRotatingSectionsGradeableIDS() {
-        $this->course_db->query("SELECT g_id FROM gradeable WHERE g_grade_by_registration = 'f' ORDER BY g_grade_start_date ASC");
-        return $this->course_db->rows();
-    }
-
     public function getGradeablesPastAndSection() {
         $this->course_db->query("
   SELECT
@@ -870,16 +581,6 @@ ORDER BY g.sections_rotating_id, g.user_id", $params);
         return $this->course_db->rows();
     }
 
-    public function getCountUsersRotatingSections() {
-        $this->course_db->query("
-SELECT rotating_section, count(*) as count
-FROM users
-WHERE (registration_section IS NOT NULL OR manual_registration)
-GROUP BY rotating_section
-ORDER BY rotating_section");
-        return $this->course_db->rows();
-    }
-
     public function getGradersForAllRotatingSections($gradeable_id) {
         $this->course_db->query("
     SELECT 
@@ -894,274 +595,6 @@ ORDER BY rotating_section");
         u.user_id
     ",array($gradeable_id));
         return $this->course_db->rows();
-    }
-
-    public function getGradersFromUserType($user_type) {
-        $this->course_db->query("SELECT user_id FROM users WHERE user_group=? ORDER BY user_id ASC", array($user_type));
-        return $this->course_db->rows();
-    }
-
-    public function getCountNullUsersRotatingSections() {
-        $this->course_db->query("
-SELECT rotating_section, count(*) as count
-FROM users
-WHERE (registration_section IS NULL and NOT manual_registration) AND rotating_section IS NOT NULL
-GROUP BY rotating_section
-ORDER BY rotating_section");
-        return $this->course_db->rows();
-    }
-
-    public function getRegisteredUserIdsWithNullRotating() {
-        $this->course_db->query("
-SELECT user_id 
-FROM users 
-WHERE
-    (rotating_section IS NULL) and 
-    (registration_section IS NOT NULL or manual_registration)
-ORDER BY user_id ASC");
-        return array_map(function($elem) { return $elem['user_id']; }, $this->course_db->rows());
-    }
-
-    public function getRegisteredUserIds() {
-        $this->course_db->query("
-SELECT user_id 
-FROM users 
-WHERE
-    (registration_section IS NOT NULL) OR 
-    (manual_registration)
-ORDER BY user_id ASC");
-        return array_map(function($elem) { return $elem['user_id']; }, $this->course_db->rows());
-    }
-
-    public function setAllUsersRotatingSectionNull() {
-        $this->course_db->query("UPDATE users SET rotating_section=NULL");
-    }
-
-    public function setNonRegisteredUsersRotatingSectionNull() {
-        $this->course_db->query("UPDATE users SET rotating_section=NULL WHERE registration_section IS NULL AND NOT manual_registration");
-    }
-
-    public function deleteAllRotatingSections() {
-        $this->course_db->query("DELETE FROM sections_rotating");
-    }
-
-    public function getMaxRotatingSection() {
-        $this->course_db->query("SELECT MAX(sections_rotating_id) as max FROM sections_rotating");
-        $row = $this->course_db->row();
-        return $row['max'];
-    }
-
-    public function getNumberRotatingSections() {
-        $this->course_db->query("SELECT COUNT(*) AS cnt FROM sections_rotating");
-        return $this->course_db->row()['cnt'];
-    }
-
-    public function insertNewRotatingSection($section) {
-        $this->course_db->query("INSERT INTO sections_rotating (sections_rotating_id) VALUES(?)", array($section));
-    }
-
-    public function setupRotatingSections($graders, $gradeable_id) {
-        $this->course_db->query("DELETE FROM grading_rotating WHERE g_id=?", array($gradeable_id));
-        foreach ($graders as $grader=>$sections){
-            foreach($sections as $i=>$section){
-                $this->course_db->query("INSERT INTO grading_rotating(g_id, user_id, sections_rotating_id) VALUES(?,?,?)", array($gradeable_id,$grader,$section));
-            }
-        }
-    }
-
-    public function updateUsersRotatingSection($section, $users) {
-        $update_array = array_merge(array($section), $users);
-        $update_string = implode(",", array_pad(array(), count($users), "?"));
-        $this->course_db->query("UPDATE users SET rotating_section=? WHERE user_id IN ({$update_string})", $update_array);
-    }
-
-    public function insertVersionDetails($g_id, $user_id, $team_id, $version, $timestamp) {
-        $this->course_db->query("
-INSERT INTO electronic_gradeable_data 
-(g_id, user_id, team_id, g_version, autograding_non_hidden_non_extra_credit, autograding_non_hidden_extra_credit, 
-autograding_hidden_non_extra_credit, autograding_hidden_extra_credit, submission_time) 
-
-VALUES(?, ?, ?, ?, 0, 0, 0, 0, ?)", array($g_id, $user_id, $team_id, $version, $timestamp));
-        if ($user_id === null) {
-            $this->course_db->query("SELECT * FROM electronic_gradeable_version WHERE g_id=? AND team_id=?",
-            array($g_id, $team_id));
-        }
-        else {
-            $this->course_db->query("SELECT * FROM electronic_gradeable_version WHERE g_id=? AND user_id=?",
-            array($g_id, $user_id));
-        }
-        $row = $this->course_db->row();
-        if (!empty($row)) {
-            $this->updateActiveVersion($g_id, $user_id, $team_id, $version);
-        }
-        else {
-            $this->course_db->query("INSERT INTO electronic_gradeable_version (g_id, user_id, team_id, active_version) VALUES(?, ?, ?, ?)",
-                array($g_id, $user_id, $team_id, $version));
-        }
-    }
-
-    public function updateActiveVersion($g_id, $user_id, $team_id, $version) {
-        if ($user_id === null) {
-            $this->course_db->query("UPDATE electronic_gradeable_version SET active_version=? WHERE g_id=? AND team_id=?",
-            array($version, $g_id, $team_id));
-        }
-        else {
-            $this->course_db->query("UPDATE electronic_gradeable_version SET active_version=? WHERE g_id=? AND user_id=?",
-            array($version, $g_id, $user_id));
-        }  
-    }
-
-    public function insertGradeableData(Gradeable $gradeable) {
-        if ($gradeable->isTeamAssignment()) {
-            $params = array($gradeable->getId(), $gradeable->getTeam()->getId(),
-                            $gradeable->getOverallComment());
-            $this->course_db->query("INSERT INTO 
-gradeable_data (g_id, gd_team_id, gd_overall_comment)
-VALUES (?, ?, ?)", $params);
-        }
-        else {
-            $params = array($gradeable->getId(), $gradeable->getUser()->getId(),
-                            $gradeable->getOverallComment());
-            $this->course_db->query("INSERT INTO 
-gradeable_data (g_id, gd_user_id, gd_overall_comment)
-VALUES (?, ?, ?)", $params);
-        }
-        return $this->course_db->getLastInsertId("gradeable_data_gd_id_seq");
-    }
-
-    public function updateGradeableData(Gradeable $gradeable) {
-        $params = array($gradeable->getOverallComment(), $gradeable->getGdId());
-
-        $this->course_db->query("UPDATE gradeable_data SET gd_overall_comment=? WHERE gd_id=?", $params);
-    }
-
-    public function insertGradeableComponentData($gd_id, GradeableComponent $component) {
-        $params = array($component->getId(), $gd_id, $component->getScore(), $component->getComment(), $component->getGrader()->getId(), $component->getGradedVersion(), $component->getGradeTime()->format("Y-m-d H:i:s"));
-        $this->course_db->query("
-INSERT INTO gradeable_component_data (gc_id, gd_id, gcd_score, gcd_component_comment, gcd_grader_id, gcd_graded_version, gcd_grade_time) 
-VALUES (?, ?, ?, ?, ?, ?, ?)", $params);
-    }
-
-    public function updateGradeableComponentData($gd_id, GradeableComponent $component) {
-        $params = array($component->getScore(), $component->getComment(), $component->getGrader()->getId(), $component->getGradedVersion(), $component->getGradeTime()->format("Y-m-d H:i:s"), $component->getId(), $gd_id);
-        $this->course_db->query("
-UPDATE gradeable_component_data SET gcd_score=?, gcd_component_comment=?, gcd_grader_id=?, gcd_graded_version=?, gcd_grade_time=? WHERE gc_id=? AND gd_id=?", $params);
-    }
-
-    public function deleteGradeableComponentData($gd_id, GradeableComponent $component) {
-        $params = array($component->getId(), $gd_id);
-        $this->course_db->query("
-DELETE FROM gradeable_component_data WHERE gc_id=? AND gd_id=?", $params);
-    }
-
-    public function checkGradeableComponentData($gd_id, GradeableComponent $component) {
-        $this->course_db->query("SELECT COUNT(*) as cnt FROM gradeable_component_data WHERE gc_id=? AND gd_id=?", 
-          array($component->getId(), $gd_id));
-        if ($this->course_db->row()['cnt'] == 0) {
-          return false;
-        }
-        return true;
-    }
-
-    public function deleteGradeableComponentMarkData($gd_id, $gc_id, GradeableComponentMark $mark) {
-        $params = array($gc_id, $gd_id, $mark->getId());
-        $this->course_db->query("
-DELETE FROM gradeable_component_mark_data WHERE gc_id=? AND gd_id=? AND gcm_id=?", $params);
-    }
-
-    public function getDataFromGCMD($gc_id, GradeableComponentMark $mark) {
-        $return_data = array();
-        $params = array($gc_id, $mark->getId());
-        $this->course_db->query("
-SELECT gd_id FROM gradeable_component_mark_data WHERE gc_id=? AND gcm_id=?", $params);
-        $rows = $this->course_db->rows();
-        foreach ($rows as $row) {
-            $this->course_db->query("
-SELECT gd_user_id FROM gradeable_data WHERE gd_id=?", array($row['gd_id']));
-            $temp_array = array();
-            $temp_array['gd_user_id'] = $this->course_db->row()['gd_user_id'];
-            $return_data[] = $temp_array;
-        }
-
-        return $return_data;
-    }
-
-    public function insertGradeableComponentMarkData($gd_id, $gc_id, GradeableComponentMark $mark) {
-        $params = array($gc_id, $gd_id, $mark->getId());
-        $this->course_db->query("
-INSERT INTO gradeable_component_mark_data (gc_id, gd_id, gcm_id)
-VALUES (?, ?, ?)", $params);
-    }
-
-    public function createNewGradeable(Gradeable $gradeable) {
-        $params = array($gradeable->getId(), $gradeable->getName(), $gradeable->getInstructionsUrl(), $gradeable->getTaInstructions(), $gradeable->getType(), var_export($gradeable->getGradeByRegistration(), true), $gradeable->getTaViewDate()->format('Y/m/d H:i:s'), $gradeable->getGradeStartDate()->format('Y/m/d H:i:s'), $gradeable->getGradeReleasedDate()->format('Y/m/d H:i:s'), $gradeable->getMinimumGradingGroup(), $gradeable->getBucket());
-        $this->course_db->query("
-INSERT INTO gradeable(g_id, g_title, g_instructions_url,g_overall_ta_instructions, g_gradeable_type, g_grade_by_registration, g_ta_view_start_date, g_grade_start_date,  g_grade_released_date,  g_min_grading_group, g_syllabus_bucket) 
-VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", $params);
-        if ($gradeable->getType() === GradeableType::ELECTRONIC_FILE) {
-            $params = array($gradeable->getId(), $gradeable->getOpenDate()->format('Y/m/d H:i:s'), $gradeable->getDueDate()->format('Y/m/d H:i:s'), var_export($gradeable->getIsRepository(), true), $gradeable->getSubdirectory(), var_export($gradeable->getTeamAssignment(),true), $gradeable->getMaxTeamSize(), $gradeable->getTeamLockDate()->format('Y/m/d H:i:s'), var_export($gradeable->getTaGrading(), true), var_export($gradeable->getStudentView(), true), var_export($gradeable->getStudentSubmit(), true),  var_export($gradeable->getStudentDownload(), true), var_export($gradeable->getStudentAnyVersion(), true), $gradeable->getConfigPath(), $gradeable->getLateDays(), $gradeable->getPointPrecision(), var_export($gradeable->getPeerGrading(), true), $gradeable->getPeerGradeSet());
-            $this->course_db->query("
-INSERT INTO electronic_gradeable(g_id, eg_submission_open_date, eg_submission_due_date, eg_is_repository, 
-eg_subdirectory, eg_team_assignment, eg_max_team_size, eg_team_lock_date, eg_use_ta_grading, eg_student_view, eg_student_submit, eg_student_download,
-eg_student_any_version, eg_config_path, eg_late_days, eg_precision, eg_peer_grading, eg_peer_grade_set) 
-VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", $params);
-        }
-    }
-
-    public function updateGradeable(Gradeable $gradeable) {
-        $params = array($gradeable->getName(), $gradeable->getInstructionsUrl(), $gradeable->getTaInstructions(), $gradeable->getType(), var_export($gradeable->getGradeByRegistration(), true), $gradeable->getTaViewDate()->format('Y/m/d H:i:s'), $gradeable->getGradeStartDate()->format('Y/m/d H:i:s'), $gradeable->getGradeReleasedDate()->format('Y/m/d H:i:s'), $gradeable->getMinimumGradingGroup(), $gradeable->getBucket(), $gradeable->getId());
-        $this->course_db->query("
-UPDATE gradeable SET g_title=?, g_instructions_url=?, g_overall_ta_instructions=?,
-g_gradeable_type=?, g_grade_by_registration=?, g_ta_view_start_date=?, g_grade_start_date=?, 
-g_grade_released_date=?, g_min_grading_group=?, g_syllabus_bucket=? WHERE g_id=?", $params);
-        if ($gradeable->getType() === 0) {
-          $params = array($gradeable->getOpenDate()->format('Y/m/d H:i:s'), $gradeable->getDueDate()->format('Y/m/d H:i:s'), var_export($gradeable->getIsRepository(), true), $gradeable->getSubdirectory(), var_export($gradeable->getTeamAssignment(),true), $gradeable->getMaxTeamSize(), $gradeable->getTeamLockDate()->format('Y/m/d H:i:s'), var_export($gradeable->getTaGrading(), true), var_export($gradeable->getStudentView(), true), var_export($gradeable->getStudentSubmit(), true), var_export($gradeable->getStudentDownload(), true), var_export($gradeable->getStudentAnyVersion(), true), $gradeable->getConfigPath(), $gradeable->getLateDays(), $gradeable->getPointPrecision(), var_export($gradeable->getPeerGrading(), true), $gradeable->getPeerGradeSet(), $gradeable->getId());
-          $this->course_db->query("
-UPDATE electronic_gradeable SET eg_submission_open_date=?, eg_submission_due_date=?, eg_is_repository=?, 
-eg_subdirectory=?, eg_team_assignment=?, eg_max_team_size=?, eg_team_lock_date=?, eg_use_ta_grading=?, eg_student_view=?, eg_student_submit=?,
-eg_student_download=?, eg_student_any_version=?, eg_config_path=?, eg_late_days=?, eg_precision=?, eg_peer_grading=?, eg_peer_grade_set=? WHERE g_id=?", $params);
-        }
-    }
-
-    public function createNewGradeableComponent(GradeableComponent $component, Gradeable $gradeable) {
-        $params = array($gradeable->getId(), $component->getTitle(), $component->getTaComment(), $component->getStudentComment(), $component->getLowerClamp(), $component->getDefault(), $component->getMaxValue(), $component->getUpperClamp(), var_export($component->getIsText(), true), $component->getOrder(), var_export($component->getIsPeer(), true), $component->getPage());
-        $this->course_db->query("
-INSERT INTO gradeable_component(g_id, gc_title, gc_ta_comment, gc_student_comment, gc_lower_clamp, gc_default, gc_max_value, gc_upper_clamp, 
-gc_is_text, gc_order, gc_is_peer, gc_page) 
-VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", $params);
-    }   
-
-    public function updateGradeableComponent(GradeableComponent $component) {
-        $params = array($component->getTitle(), $component->getTaComment(), $component->getStudentComment(), $component->getLowerClamp(), $component->getDefault(), $component->getMaxValue(), $component->getUpperClamp(), var_export($component->getIsText(), true), $component->getOrder(), var_export($component->getIsPeer(), true), $component->getPage(), $component->getId());
-        $this->course_db->query("
-UPDATE gradeable_component SET gc_title=?, gc_ta_comment=?, gc_student_comment=?, gc_lower_clamp=?, gc_default=?, gc_max_value=?, gc_upper_clamp=?, gc_is_text=?, gc_order=?, gc_is_peer=?, gc_page=? WHERE gc_id=?", $params);
-    }
-
-    public function deleteGradeableComponent(GradeableComponent $component) {
-        $this->course_db->query("DELETE FROM gradeable_component_data WHERE gc_id=?",array($component->getId()));
-        $this->course_db->query("DELETE FROM gradeable_component WHERE gc_id=?", array($component->getId()));
-    }
-
-    public function createGradeableComponentMark(GradeableComponentMark $mark) {
-        $params = array($mark->getGcId(), $mark->getPoints(), $mark->getNoteNoDecode(), $mark->getOrder());
-
-        $this->course_db->query("
-INSERT INTO gradeable_component_mark (gc_id, gcm_points, gcm_note, gcm_order)
-VALUES (?, ?, ?, ?)", $params);
-        return $this->course_db->getLastInsertId();
-    }
-
-    public function updateGradeableComponentMark(GradeableComponentMark $mark) {
-        $params = array($mark->getGcId(), $mark->getPoints(), $mark->getNoteNoDecode(), $mark->getOrder(), $mark->getId());
-
-        $this->course_db->query("
-UPDATE gradeable_component_mark SET gc_id=?, gcm_points=?, gcm_note=?, gcm_order=?
-WHERE gcm_id=?", $params);
-    }
-
-    public function deleteGradeableComponentMark(GradeableComponentMark $mark) {
-        $this->course_db->query("DELETE FROM gradeable_component_mark_data WHERE gcm_id=?",array($mark->getId()));
-        $this->course_db->query("DELETE FROM gradeable_component_mark WHERE gcm_id=?", array($mark->getId()));
     }
 
     public function getGradeableData($gradeable_id) {
@@ -1233,310 +666,6 @@ WHERE gcm_id=?", $params);
         }
 
         return $data;
-    }
-
-    public function updateUserViewedDate(Gradeable $gradeable) {
-        if ($gradeable->getGdId() !== null) {
-            $this->course_db->query("UPDATE gradeable_data SET gd_user_viewed_date = NOW() WHERE gd_id=?",
-                array($gradeable->getGdId()));
-        }
-    }
-
-    public function getSession($session_id) {
-        $this->submitty_db->query("SELECT * FROM sessions WHERE session_id=?", array($session_id));
-        return $this->submitty_db->row();
-    }
-
-    public function newSession($session_id, $user_id, $csrf_token) {
-        $this->submitty_db->query("INSERT INTO sessions (session_id, user_id, csrf_token, session_expires) 
-                                   VALUES(?,?,?,current_timestamp + interval '336 hours')",
-            array($session_id, $user_id, $csrf_token));
-
-    }
-
-    public function updateSessionExpiration($session_id) {
-        $this->submitty_db->query("UPDATE sessions SET session_expires=(current_timestamp + interval '336 hours') 
-                                   WHERE session_id=?", array($session_id));
-    }
-
-    public function removeExpiredSessions() {
-        $this->submitty_db->query("DELETE FROM sessions WHERE session_expires < current_timestamp");
-    }
-
-    public function removeSessionById($session_id) {
-        $this->submitty_db->query("DELETE FROM sessions WHERE session_id=?", array($session_id));
-    }
-    
-    public function getAllGradeablesIdsAndTitles() {
-        $this->course_db->query("SELECT g_id, g_title FROM gradeable ORDER BY g_title ASC");
-        return $this->course_db->rows();
-    }
-
-    public function getAllGradeablesIds() {
-        $this->course_db->query("SELECT g_id FROM gradeable ORDER BY g_id");
-        return $this->course_db->rows();
-    }
-    public function getAllElectronicGradeablesIds() {
-        $this->course_db->query("SELECT g_id, g_title FROM gradeable WHERE g_gradeable_type=0 ORDER BY g_grade_released_date DESC");
-        return $this->course_db->rows();
-    }
-      
-    public function createTeam($g_id, $user_id, $registration_section, $rotating_section) {
-        $this->course_db->query("SELECT COUNT(*) AS cnt FROM gradeable_teams");
-        $team_id_prefix = strval($this->course_db->row()['cnt']);
-        if (strlen($team_id_prefix) < 5) $team_id_prefix = str_repeat("0", 5-strlen($team_id_prefix)) . $team_id_prefix;
-        $team_id = "{$team_id_prefix}_{$user_id}";
-        
-        $params = array($team_id, $g_id, $registration_section, $rotating_section);
-        $this->course_db->query("INSERT INTO gradeable_teams (team_id, g_id, registration_section, rotating_section) VALUES(?,?,?,?)", $params);
-        $this->course_db->query("INSERT INTO teams (team_id, user_id, state) VALUES(?,?,1)", array($team_id, $user_id));
-        return $team_id;
-    }
-
-    public function updateTeamRegistrationSection($team_id, $section) {
-        $this->course_db->query("UPDATE gradeable_teams SET registration_section=? WHERE team_id=?", array($section, $team_id));
-    }
-
-    public function updateTeamRotatingSection($team_id, $section) {
-        $this->course_db->query("UPDATE gradeable_teams SET rotating_section=? WHERE team_id=?", array($section, $team_id));
-    }
-
-    public function leaveTeam($team_id, $user_id) {
-        $this->course_db->query("DELETE FROM teams AS t
-          WHERE team_id=? AND user_id=? AND state=1", array($team_id, $user_id));
-    }
-
-    public function sendTeamInvitation($team_id, $user_id) {
-        $this->course_db->query("INSERT INTO teams (team_id, user_id, state) VALUES(?,?,0)", array($team_id, $user_id));
-    }
-
-    public function acceptTeamInvitation($team_id, $user_id) {
-        $this->course_db->query("INSERT INTO teams (team_id, user_id, state) VALUES(?,?,1)", array($team_id, $user_id));
-    }
-
-    public function cancelTeamInvitation($team_id, $user_id) {
-        $this->course_db->query("DELETE FROM teams WHERE team_id=? AND user_id=? AND state=0", array($team_id, $user_id));
-    }
-
-    public function declineAllTeamInvitations($g_id, $user_id) {
-        $this->course_db->query("DELETE FROM teams AS t USING gradeable_teams AS gt
-          WHERE gt.g_id=? AND gt.team_id = t.team_id AND t.user_id=? AND t.state=0", array($g_id, $user_id));
-    }
-
-    public function getTeamById($team_id) {
-        $this->course_db->query("
-          SELECT team_id, registration_section, rotating_section
-          FROM gradeable_teams
-          WHERE team_id=?",
-          array($team_id));
-        if (count($this->course_db->rows()) === 0) {
-            return null;
-        }
-        $details = $this->course_db->row();
-
-        $this->course_db->query("SELECT user_id, state FROM teams WHERE team_id=? ORDER BY user_id", array($team_id));
-        $details['users'] = $this->course_db->rows();
-        return new Team($this->core, $details);
-    }
-
-    public function getTeamByGradeableAndUser($g_id, $user_id) {
-        $this->course_db->query("
-          SELECT team_id, registration_section, rotating_section
-          FROM gradeable_teams
-          WHERE g_id=? AND team_id IN (
-            SELECT team_id
-            FROM teams
-            WHERE user_id=? AND state=1)",
-          array($g_id, $user_id));
-        if (count($this->course_db->rows()) === 0) {
-            return null;
-        }
-        $details = $this->course_db->row();
-
-        $this->course_db->query("SELECT user_id, state FROM teams WHERE team_id=? ORDER BY user_id", array($details['team_id']));
-        $details['users'] = $this->course_db->rows();
-        return new Team($this->core, $details);
-    }
-
-    public function getTeamsByGradeableId($g_id) {
-        $this->course_db->query("
-          SELECT team_id, registration_section, rotating_section
-          FROM gradeable_teams
-          WHERE g_id=?
-          ORDER BY team_id",
-          array($g_id));
-
-        $all_teams_details = array();
-        foreach($this->course_db->rows() as $row) {
-            $all_teams_details[$row['team_id']] = $row;
-        }
-
-        $teams = array();
-        foreach($all_teams_details as $team_id => $details) {
-            $this->course_db->query("SELECT user_id, state FROM teams WHERE team_id=? ORDER BY user_id", array($team_id));
-            $details['users'] = $this->course_db->rows();
-            $teams[] = new Team($this->core, $details);
-        }
-
-        return $teams;
-    }
-
-    public function getTotalTeamCountByGradingSections($g_id, $sections, $section_key) {
-        $return = array();
-        $params = array($g_id);
-        $sections_query = "";
-        if (count($sections) > 0) {
-            $sections_query = "{$section_key} IN (".implode(",", array_fill(0, count($sections), "?")).") AND";
-            $params = array_merge($sections, $params);
-        }
-        $this->course_db->query("
-SELECT count(*) as cnt, {$section_key} 
-FROM gradeable_teams 
-WHERE {$sections_query} g_id=? AND team_id IN (
-  SELECT team_id
-  FROM teams
-)
-GROUP BY {$section_key} 
-ORDER BY {$section_key}", $params);
-        foreach ($this->course_db->rows() as $row) {
-            $return[$row[$section_key]] = intval($row['cnt']);
-        }
-        foreach ($sections as $section) {
-            if (!isset($return[$section])) $return[$section] = 0;
-        }
-        ksort($return);
-        return $return;
-    }
-
-    public function getUsersWithoutTeamByGradingSections($g_id, $sections, $section_key) {
-        $return = array();
-        $params = array($g_id);
-        $sections_query = "";
-        if (count($sections) > 0) {
-            $sections_query= "{$section_key} IN (".implode(",", array_fill(0, count($sections), "?")).") AND";
-            $params = array_merge($sections, $params);
-        }
-        $this->course_db->query("
-SELECT count(*) as cnt, {$section_key} 
-FROM users 
-WHERE {$sections_query} user_id NOT IN (
-  SELECT user_id
-  FROM gradeable_teams NATURAL JOIN teams
-  WHERE g_id=?
-  ORDER BY user_id
-)
-GROUP BY {$section_key} 
-ORDER BY {$section_key}", $params);
-        foreach ($this->course_db->rows() as $row) {
-            $return[$row[$section_key]] = intval($row['cnt']);
-        }
-        foreach ($sections as $section) {
-            if (!isset($return[$section])) $return[$section] = 0;
-        }
-        ksort($return);
-        return $return;
-    }
-
-    public function getGradedComponentsCountByTeamGradingSections($g_id, $sections, $section_key) {
-        $return = array();
-        $params = array($g_id);
-        $where = "";
-        if (count($sections) > 0) {
-            $where = "WHERE {$section_key} IN (".implode(",", array_fill(0, count($sections), "?")).")";
-            $params = array_merge($params, $sections);
-        }
-        $this->course_db->query("
-SELECT count(gt.*) as cnt, gt.{$section_key}
-FROM gradeable_teams AS gt
-INNER JOIN (
-  SELECT * FROM gradeable_data AS gd LEFT JOIN gradeable_component_data AS gcd ON gcd.gd_id = gd.gd_id WHERE g_id=?
-) AS gd ON gt.team_id = gd.gd_team_id
-{$where}
-GROUP BY gt.{$section_key}
-ORDER BY gt.{$section_key}", $params);
-        foreach ($this->course_db->rows() as $row) {
-            $return[$row[$section_key]] = intval($row['cnt']);
-        }
-        return $return;
-    }
-
-    public function getUsersWithLateDays() {
-      $this->course_db->query("
-        SELECT u.user_id, user_firstname, user_preferred_firstname, 
-          user_lastname, allowed_late_days, since_timestamp::timestamp::date
-        FROM users AS u
-        FULL OUTER JOIN late_days AS l
-          ON u.user_id=l.user_id
-        WHERE allowed_late_days IS NOT NULL
-          AND allowed_late_days>0
-        ORDER BY
-          user_email ASC, since_timestamp DESC;");
-
-      $return = array();
-      foreach($this->course_db->rows() as $row){
-        $return[] = new SimpleLateUser($this->core, $row);
-      }
-      return $return;
-    }
-
-    public function getUsersWithExtensions($gradeable_id) {
-      $this->course_db->query("
-        SELECT u.user_id, user_firstname,
-          user_preferred_firstname, user_lastname, late_day_exceptions
-        FROM users as u
-        FULL OUTER JOIN late_day_exceptions as l
-          ON u.user_id=l.user_id
-        WHERE g_id=?
-          AND late_day_exceptions IS NOT NULL
-          AND late_day_exceptions>0
-        ORDER BY user_email ASC;", array($gradeable_id));
-
-      $return = array();
-      foreach($this->course_db->rows() as $row){
-        $return[] = new SimpleLateUser($this->core, $row);
-      }
-      return $return;
-    }
-
-    public function updateLateDays($user_id, $timestamp, $days){
-        $this->course_db->query("
-          UPDATE late_days
-          SET allowed_late_days=?
-          WHERE user_id=?
-            AND since_timestamp=?", array($days, $user_id, $timestamp));
-        if(count($this->course_db->rows())==0){
-          $this->course_db->query("
-            INSERT INTO late_days
-            (user_id, since_timestamp, allowed_late_days)
-            VALUES(?,?,?)", array($user_id, $timestamp, $days));
-        }
-    }
-
-    public function updateExtensions($user_id, $g_id, $days){
-        $this->course_db->query("
-          UPDATE late_day_exceptions
-          SET late_day_exceptions=?
-          WHERE user_id=?
-            AND g_id=?;", array($days, $user_id, $g_id));
-        if(count($this->course_db->rows())==0){
-          $this->course_db->query("
-            INSERT INTO late_day_exceptions
-            (user_id, g_id, late_day_exceptions)
-            VALUES(?,?,?)", array($user_id, $g_id, $days));
-        }
-    }
-    
-    public function getPeerGradingAssignNumber($gradeable_id) {
-        $this->course_db->query("SELECT eg_peer_grade_set FROM electronic_gradeable WHERE g_id=?", array($gradeable_id));
-        return $this->course_db->row()['eg_peer_grade_set'];
-    }
-    
-    public function clearPeerGradingAssignments($gradeable_id) {
-        $this->course_db->query("DELETE FROM peer_assign WHERE g_id=?", array($gradeable_id));
-    }
-    
-    public function insertPeerGradingAssignment($grader, $student, $gradeable_id) {
-        $this->course_db->query("INSERT INTO peer_assign(grader_id, user_id, g_id) VALUES (?,?,?)", array($grader, $student, $gradeable_id));
     }
 }
